requirements.txt
requests
beautifulsoup4
yfinance
pandas
vaderSentiment
python-dotenv
